+++ tests/test_image/common.py
@@ -230,56 +230,51 @@ class TestSetSize_All:
         type(self).v_image = ImageClass.from_file("tests/images/vert.jpg")
 
     def test_auto(self):
-        max_width = ImageClass._pixels_cols(cols=columns)
-        max_height = ImageClass._pixels_lines(lines=lines - 2)
+        frame_width_px = ImageClass._pixels_cols(cols=columns)
+        frame_height_px = ImageClass._pixels_lines(lines=lines - 2)
 
         _original_size = self.image.original_size
         try:
-            # Equal to max size
-            self.image._original_size = (max_width, max_height)
+            # Equal to frame size
+            self.image._original_size = (frame_width_px, frame_height_px)
             self.image.set_size(Size.AUTO)
             assert self.image._valid_size(Size.ORIGINAL) == self.image.size
-            assert not self.image._fit_to_width
 
-            # Horizontally larger than max size
-            self.image._original_size = (max_width + 20, max_height - 20)
+            # Horizontally larger than frame size
+            self.image._original_size = (frame_width_px + 20, frame_height_px - 20)
             self.image.set_size(Size.AUTO)
             assert (
                 self.image._valid_size(Size.ORIGINAL)
                 != self.image.size
                 == self.image._valid_size(Size.FIT)
             )
-            assert not self.image._fit_to_width
 
-            # Vertically larger than max size
-            self.image._original_size = (max_width - 20, max_height + 20)
+            # Vertically larger than frame size
+            self.image._original_size = (frame_width_px - 20, frame_height_px + 20)
             self.image.set_size(Size.AUTO)
             assert (
                 self.image._valid_size(Size.ORIGINAL)
                 != self.image.size
                 == self.image._valid_size(Size.FIT)
             )
-            assert not self.image._fit_to_width
 
-            # Horizontally and vertically larger than max size
-            self.image._original_size = (max_width + 20, max_height + 20)
+            # Horizontally and vertically larger than frame size
+            self.image._original_size = (frame_width_px + 20, frame_height_px + 20)
             self.image.set_size(Size.AUTO)
             assert (
                 self.image._valid_size(Size.ORIGINAL)
                 != self.image.size
                 == self.image._valid_size(Size.FIT)
             )
-            assert not self.image._fit_to_width
 
-            # Smaller than max size
-            self.image._original_size = (max_width - 20, max_height - 20)
+            # Smaller than frame size
+            self.image._original_size = (frame_width_px - 20, frame_height_px - 20)
             self.image.set_size(Size.AUTO)
             assert (
                 self.image._valid_size(Size.ORIGINAL)
                 == self.image.size
                 != self.image._valid_size(Size.FIT)
             )
-            assert not self.image._fit_to_width
         finally:
             self.image._original_size = _original_size
 
@@ -289,19 +284,15 @@ class TestSetSize_All:
 
         self.image.set_size(width=Size.FIT)
         assert self.image.size == size
-        assert not self.image._fit_to_width
 
         self.image.set_size(height=Size.FIT)
         assert self.image.size == size
-        assert not self.image._fit_to_width
 
         self.image.set_size(width=None)
         assert self.image.size == size
-        assert not self.image._fit_to_width
 
         self.image.set_size(height=None)
         assert self.image.size == size
-        assert not self.image._fit_to_width
 
     # a PASS is valid only if the previous test passed
     def test_fit(self):
@@ -327,7 +318,6 @@ class TestSetSize_All:
         self.image.set_size(width=Size.FIT_TO_WIDTH)
         assert self.image.width == columns
         assert proportional(self.image)
-        assert self.image._fit_to_width
 
         self.h_image.set_size(width=Size.FIT_TO_WIDTH)
         assert self.h_image.width == columns
@@ -341,7 +331,6 @@ class TestSetSize_All:
         self.image.set_size(height=Size.FIT_TO_WIDTH)
         assert self.image.width == columns
         assert proportional(self.image)
-        assert self.image._fit_to_width
 
         self.h_image.set_size(height=Size.FIT_TO_WIDTH)
         assert self.h_image.width == columns
@@ -355,7 +344,6 @@ class TestSetSize_All:
         self.image.set_size(width=_size)
         assert self.image.width == _size
         assert proportional(self.image)
-        assert not self.image._fit_to_width
 
         self.h_image.set_size(width=_size)
         assert self.h_image.width == _size
@@ -369,7 +357,6 @@ class TestSetSize_All:
         self.image.set_size(height=_size)
         assert self.image.height == _size
         assert proportional(self.image)
-        assert not self.image._fit_to_width
 
         self.h_image.set_size(height=_size)
         assert self.h_image.height == _size
@@ -384,7 +371,6 @@ class TestSetSize_All:
         self.image.set_size(width=Size.ORIGINAL)
         assert self.image.width == ImageClass._pixels_cols(pixels=ori_width)
         assert proportional(self.image)
-        assert not self.image._fit_to_width
 
         ori_width, ori_height = self.h_image.original_size
         self.h_image.set_size(width=Size.ORIGINAL)
@@ -401,7 +387,6 @@ class TestSetSize_All:
         self.image.set_size(height=Size.ORIGINAL)
         assert self.image.height == ImageClass._pixels_lines(pixels=ori_height)
         assert proportional(self.image)
-        assert not self.image._fit_to_width
 
         ori_width, ori_height = self.h_image.original_size
         self.h_image.set_size(height=Size.ORIGINAL)
@@ -413,19 +398,6 @@ class TestSetSize_All:
         assert self.v_image.height == ImageClass._pixels_lines(pixels=ori_height)
         assert proportional(self.v_image)
 
-    def test_allowance(self):
-        self.h_image.set_size()
-        assert self.h_image.width == columns
-
-        self.v_image.set_size()
-        assert self.v_image.height == lines - 2
-
-        self.h_image.set_size(h_allow=2)
-        assert self.h_image.width == columns - 2
-
-        self.v_image.set_size(v_allow=3)
-        assert self.v_image.height == lines - 3
-
     def test_can_exceed_terminal_size(self):
         self.image.set_size(width=columns + 1)
         assert self.image.width == columns + 1
@@ -434,27 +406,6 @@ class TestSetSize_All:
         self.image.set_size(height=lines + 1)
         assert self.image.height == lines + 1
         assert proportional(self.image)
-
-    # This test passes only when the cell ratio is about 0.5
-    def test_maxsize(self):
-        self.image.set_size(maxsize=(100, 50))
-        assert self.image.size == (100, 50)
-        self.image.set_size(maxsize=(100, 55))
-        assert self.image.size == (100, 50)
-        self.image.set_size(maxsize=(110, 50))
-        assert self.image.size == (100, 50)
-
-        self.image.set_size(width=100, maxsize=(200, 100))
-        assert self.image.size == (100, 50)
-        self.image.set_size(height=50, maxsize=(200, 100))
-        assert self.image.size == (100, 50)
-
-    # This test passes only when the cell ratio is about 0.5
-    def test_maxsize_allowance_ignore(self):
-        self.image.set_size(h_allow=2, v_allow=3, maxsize=(100, 50))
-        assert self.image.size == (100, 50)
-        self.image.set_size(h_allow=2, v_allow=3, maxsize=(100, 50))
-        assert self.image.size == (100, 50)
 
 
 class TestCellRatio_Text: